Update to Minecraft 1.19.2
Cleanup code and minimize static abuse
Add error messages for invalid client IDs, secrets, and channel name instead of sending a stack trace.
Fix issues when Vault was not on the server.
Fix errors when attempting to add or remove groups that didn't exist (a clean error is now sent in console)